+++ src/orchard/meta.clj
@@ -133,9 +133,9 @@
                                   demunged)
         [ns-str name-str] (->> (str/split wo-method #"/")
                                (remove #(re-matches #"fn|eval\d+" %)))
-        ns (some-> ns-str symbol find-ns)
+        ns (some-> ns-str not-empty symbol find-ns)
         resolved (when (and ns name-str)
-                   (ns-resolve ns (symbol name-str)))]
+                   (some->> name-str not-empty symbol (ns-resolve ns)))]
     (when (var? resolved)
       resolved)))