+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## master (unreleased)
 
 - [#362](https://github.com/clojure-emacs/orchard/pull/362): Info: don't crash if symbol contains $ or /.
-- [#365](https://github.com/clojure-emacs/orchard/pull/365): Inspector: don't datafy vars.
 - [#363](https://github.com/clojure-emacs/orchard/pull/363): Inspector: show duplicates in analytics.
+- [#365](https://github.com/clojure-emacs/orchard/pull/365): Inspector: don't datafy vars.
+- [#367](https://github.com/clojure-emacs/orchard/pull/367): Inspector: pretty-print arrays distinctively from vectors.
 
 ## 0.37.1 (2025-10-14)